requirements.txt
kivy
kivymd
rembg
onnxruntime
pillow
plyer
